# public/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Sportify",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:3000"
		}
	],
	"paths": {
		"/adminAccounts/list": {
			"get": {
				"tags": ["admins"],
				"summary": "getAdminsList",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer"
						},
						"description": "used for pagination"
					},
					{
						"name": "recordsPerPage",
						"in": "query",
						"schema": {
							"type": "integer"
						},
						"description": "number of records to include in each page"
					},
					{
						"name": "searchQuery",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "text string used for filtering"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Admin"
									}
								}
							}
						}
					}
				}
			}
		},
		"/adminAccounts": {
			"get": {
				"tags": ["admins"],
				"summary": "getAdmins",
				"parameters": [
					{
						"name": "adminEmails",
						"in": "query",
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							},
							"maxItems": 1000
						},
						"description": "An array of Emails unique to the Admins (max 1000)"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Admin"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["admins"],
				"summary": "createAdminAccount",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string"
									},
									"firstName": {
										"type": "string"
									},
									"lastName": {
										"type": "string"
									}
								},
								"required": ["email", "firstName", "lastName"]
							},
							"example": {
								"email": "admin@aucegypt.edu",
								"firstName": "John",
								"lastName": "Doe"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"example": {
									"message": "Successful Response",
									"admin": {
										"email": "admin@example.com",
										"firstName": "John",
										"lastName": "Doe"
									}
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["admins"],
				"summary": "updateAdminAccount",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"example": "admin@aucegypt.edu"
									},
									"firstName": {
										"type": "string",
										"example": "John"
									},
									"lastName": {
										"type": "string",
										"example": "Doe"
									}
								},
								"required": ["email"]
							},
							"example": {
								"email": "admin@aucegypt.edu",
								"firstName": "John",
								"lastName": "Doe"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Admin"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["admins"],
				"summary": "deleteAdminAccount",
				"requestBody": {
					"description": "List of emails to delete",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"adminEmails": {
										"type": "array",
										"items": {
											"type": "string",
											"format": "email"
										}
									}
								},
								"required": ["adminEmails"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"enum": ["Success", "Fail"]
								}
							}
						}
					}
				}
			}
		},
		"/courts/list": {
			"get": {
				"tags": ["courts"],
				"summary": "getCourtsList",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer"
						},
						"description": "used for pagination"
					},
					{
						"name": "recordsPerPage",
						"in": "query",
						"schema": {
							"type": "integer"
						},
						"description": "number of records to include in each page"
					},
					{
						"name": "searchQuery",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "text string used for filtering"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Court"
										},
										{
											"type": "integer"
										}
									]
								}
							}
						}
					}
				}
			}
		},
		"/courts": {
			"get": {
				"tags": ["courts"],
				"summary": "getCourts",
				"parameters": [
					{
						"name": "courtIds",
						"in": "query",
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							},
							"maxItems": 1000
						},
						"description": "An array of IDs unique to the courts (max 1000)"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Court"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["courts"],
				"summary": "createCourt",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string",
										"description": "Name of the court"
									},
									"sport": {
										"type": "string",
										"description": "Dedicated Sport"
									},
									"location": {
										"type": "string",
										"description": "Location of the court "
									},
									"description": {
										"type": "string",
										"description": "Description of the court (optional)"
									}
								},
								"required": ["name", "sport"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Court"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["courts"],
				"summary": "updateCourt",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"id": {
										"type": "integer",
										"description": "Id of Court to update"
									},
									"name": {
										"type": "string",
										"description": "Updated name of the court"
									},
									"location": {
										"type": "string",
										"description": "Updated location of the court"
									},
									"description": {
										"type": "string",
										"description": "Updated description of the court"
									},
									"sport": {
										"type": "string",
										"description": "Updated sport of the court"
									}
								},
								"required": ["id"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Court"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["courts"],
				"summary": "deleteCourt",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"courtIds": {
										"type": "array",
										"items": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"description": "An array of courtIds to delete"
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"enum": ["Success", "Fail"]
								}
							}
						}
					}
				}
			}
		},
		"/partyPosts/list": {
			"get": {
				"tags": ["partyPost"],
				"summary": "getPartyPostList",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer"
						},
						"description": "used for pagination"
					},
					{
						"name": "recordsPerPage",
						"in": "query",
						"schema": {
							"type": "integer"
						},
						"description": "number of records to include in each page"
					},
					{
						"name": "searchQuery",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "text string used for filtering"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/PartyPost"
									}
								}
							}
						}
					}
				}
			}
		},
		"/partyPosts": {
			"get": {
				"tags": ["partyPost"],
				"summary": "getPartyPosts",
				"parameters": [
					{
						"name": "partyIds",
						"in": "query",
						"schema": {
							"type": "integer"
						},
						"description": "list of IDs"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/PartyPost"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["partyPost"],
				"summary": "createPartyPost",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"ownerEmail": {
										"type": "string",
										"description": "Email of owner of the Event"
									},
									"eventName": {
										"type": "string",
										"description": "Name of the event"
									},
									"sport": {
										"type": "string",
										"description": "Sport of the Event"
									},
									"location": {
										"type": "string",
										"description": "Location of the Event (optional)"
									},
									"description": {
										"type": "string",
										"description": "Description of the Event (optional)"
									},
									"startTime": {
										"type": "string",
										"description": "Starting time of the Event (optional)"
									},
									"endTime": {
										"type": "string",
										"description": "Ending Time of the Event (optional)"
									}
								},
								"required": ["ownerEmail", "eventName", "sport"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PartyPost"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["partyPost"],
				"summary": "updatePartyPost",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"id": {
										"type": "integer"
									},
									"ownerEmail": {
										"type": "string",
										"description": "Email of owner of the Event"
									},
									"eventName": {
										"type": "string",
										"description": "Name of the event"
									},
									"sport": {
										"type": "string",
										"description": "Sport of the Event"
									},
									"location": {
										"type": "string",
										"description": "Location of the Event (optional)"
									},
									"description": {
										"type": "string",
										"description": "Description of the Event (optional)"
									},
									"startTime": {
										"type": "string",
										"description": "Starting time of the Event (optional)"
									},
									"endTime": {
										"type": "string",
										"description": "Ending Time of the Event (optional)"
									}
								},
								"required": ["id"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PartyPost"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["partyPost"],
				"summary": "deletePartyPost",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"partyIds": {
										"type": "array",
										"items": {
											"type": "integer"
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"enum": ["Success", "Fail"]
								}
							}
						}
					}
				}
			}
		},
		"/students/list": {
			"get": {
				"tags": ["students"],
				"summary": "getStudentsList",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer"
						},
						"description": "used for pagination"
					},
					{
						"name": "recordsPerPage",
						"in": "query",
						"schema": {
							"type": "integer"
						},
						"description": "number of records per page"
					},
					{
						"name": "searchQuery",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "text string used for filtering"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Student"
									}
								}
							}
						}
					}
				}
			}
		},
		"/students": {
			"get": {
				"tags": ["students"],
				"summary": "getStudents",
				"parameters": [
					{
						"name": "studentEmails",
						"in": "query",
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "email"
							}
						},
						"description": "Array of student emails"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Student"
									}
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["students"],
				"summary": "updateStudent",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string"
									},
									"firstName": {
										"type": "string"
									},
									"lastName": {
										"type": "string"
									},
									"photoLink": {
										"type": "string"
									},
									"isTrainer": {
										"type": "boolean"
									},
									"phoneNumber": {
										"type": "string"
									}
								},
								"required": ["email"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Student"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["students"],
				"summary": "(Un)Ban Students",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"isBanned": {
										"type": "boolean"
									},
									"studentEmails": {
										"type": "array",
										"items": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"enum": ["Success", "Fail"]
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Admin": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				}
			},
			"Court": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"id": {
						"type": "integer"
					},
					"sport": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				}
			},
			"PartyPost": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"ownerEmail": {
						"type": "string"
					},
					"member": {
						"type": "string"
					},
					"eventName": {
						"type": "string"
					},
					"sport": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startTime": {
						"type": "string",
						"format": "date-time"
					},
					"endTime": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"Student": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"photoLink": {
						"type": "string",
						"nullable": true
					},
					"isTrainer": {
						"type": "boolean"
					},
					"phoneNumber": {
						"type": "string"
					},
					"isBanned": {
						"type": "boolean"
					}
				}
			}
		}
	}
}